requirements.txt
requests
ipykernel
python-dotenv 
psycopg2-binary